Cleanup typing
